+++ R/get_nomalized_counts.R
@@ -8,7 +8,7 @@
 #'        Default is 5
 #' @param exclude_downstreamCodon An integer specifying the number of codons to exclude from the end of the coding sequence.
 #'        Default is 5.
-#' @param norm_type The nomalization methods for ribosome density. "average" is calculated by
+#' The nomalization methods for ribosome density. "average" is calculated by
 #' the count at each position divided by mean density across cds region. "rpm"
 #' is calculated by the count at each position divided by the total counts and multiplied with 10^6.
 #' Default is "average".
@@ -28,8 +28,7 @@
 setGeneric("get_nomalized_counts",
            function(object,
                     exclude_upstreamCodon = 5,
-                    exclude_downstreamCodon = 5,
-                    norm_type = c("average","rpm"),...) standardGeneric("get_nomalized_counts"))
+                    exclude_downstreamCodon = 5,...) standardGeneric("get_nomalized_counts"))
 
 
 
@@ -43,9 +42,8 @@ setMethod("get_nomalized_counts",
           signature(object = "ribosomeObj"),
           function(object,
                    exclude_upstreamCodon = 5,
-                   exclude_downstreamCodon = 5,
-                   norm_type = c("average","rpm"),...){
-            norm_type <- match.arg(norm_type,c("average","rpm"))
+                   exclude_downstreamCodon = 5,...){
+            # norm_type <- match.arg(norm_type,c("average","rpm"))
             # ============================================================================
             # at individual position was nomalized to the average counts of the transcript
             # ============================================================================
@@ -73,30 +71,28 @@ setMethod("get_nomalized_counts",
             purrr::map_df(seq_along(sp),function(x){
               tmp <- subset(qc_df,sample == sp[x])
 
-              # check normalization methods
-              if(norm_type == "average"){
-                total_exp <- tmp %>%
-                  dplyr::left_join(y = gene_lenth,by = "trans_id") %>%
-                  dplyr::filter(trans_pos >= `5UTR_length` + 1 + exclude_upstreamCodon*3 & trans_pos <= CDS_length + `5UTR_length` - exclude_downstreamCodon*3) %>%
-                  dplyr::group_by(trans_id) %>%
-                  dplyr::summarise(total_reads = sum(counts))
+              # normalization methods rpm
+              total_counts <- sum(tmp$counts)
+              tmp$rpm <- (tmp$counts/total_counts)*10^6
 
-                gene_exp <- gene_lenth %>%
-                  dplyr::filter(trans_id %in% unique(total_exp$trans_id)) %>%
-                  dplyr::left_join(y = total_exp,by = "trans_id")
+              # normalization methods average
+              total_exp <- tmp %>%
+                dplyr::left_join(y = gene_lenth,by = "trans_id") %>%
+                dplyr::filter(trans_pos >= `5UTR_length` + 1 + exclude_upstreamCodon*3 & trans_pos <= CDS_length + `5UTR_length` - exclude_downstreamCodon*3) %>%
+                dplyr::group_by(trans_id) %>%
+                dplyr::summarise(total_reads = sum(counts))
 
-                gene_exp$average_exp <- gene_exp$total_reads/gene_exp$CDS_length
+              gene_exp <- gene_lenth %>%
+                dplyr::filter(trans_id %in% unique(total_exp$trans_id)) %>%
+                dplyr::left_join(y = total_exp,by = "trans_id")
 
-                qc_df_new <- tmp %>%
-                  dplyr::left_join(y = gene_exp[,c("trans_id","average_exp")],by = "trans_id") %>%
-                  dplyr::filter(trans_id %in% unique(gene_exp$trans_id)) %>%
-                  dplyr::mutate(norm_exp = counts/average_exp) %>%
-                  dplyr::select(-average_exp)
-              }else{
-                total_counts <- sum(tmp$counts)
-                tmp$norm_exp <- (tmp$counts/total_counts)*10^6
-                qc_df_new <- tmp
-              }
+              gene_exp$average_exp <- gene_exp$total_reads/gene_exp$CDS_length
+
+              qc_df_new <- tmp %>%
+                dplyr::left_join(y = gene_exp[,c("trans_id","average_exp")],by = "trans_id") %>%
+                dplyr::filter(trans_id %in% unique(gene_exp$trans_id)) %>%
+                dplyr::mutate(average = counts/average_exp) %>%
+                dplyr::select(-average_exp)
 
               return(qc_df_new)
             }) -> qc_df_nomalized